Added fts_* implementations
Solaris does not provide fts_* functions, so copied from glibc with mods to get working on Solaris.
If the function fts_open is available on the system, then fts_sun.h and fts_sun.c become empty.
Fixed a problem with fts where only the root directory entry would be returned on solaris