Prevent skipping elements when reading CPE from within XCCDF benchmark
A much better fix would be to rework the parser but this should do.
The change is basically about catching </cpe-list> and exiting when
we encounter it. Previously we would read the element after that
and XCCDF parser would skip it and start with the element that comes
after it.