A brand new assume macro!
This one is usefull for our *_add_* and *_set_* functions,
which return true when operation passed. If you want to
assert for the result of operation in debug mode, but
you still want to commit the operation in NDEBUG, you cannot
use assert(). But you can use assume_ex() instead.

Example:
  assume_ex(xccdf_rule_add_status\(rule, status));

This should help us catch internal (and often not lethal) problems
during development while keeping users not annoyed.