Print out status of downloading.
The oval files tend to be quite big and their download may take time.
The user shall be notified about the proceedings.

Note for future generations: Once we decide to migrate oscap_acquire
module to the library and rewrite stderr fprintfs to oscap_seterror,
the printf calls shall be moved to the oscap-xccdf.c. Now, it is not
feasible as it would only mess stderr and stdout messages.