Remove a dirt (Christmas cleaning).
If You have ever wondered whether it is good idea to comment-out
some code, I have a challenge for you. Go and search for history
of these leftover and find out which of them have been later
implemented (in another place under different name), which have
been wrong from day zero, and which have not been needed so far.