Do not blindly take first policy.
This does not change the behavior at all, it only cleanups the code
by moving responsibility/abstraction to the proper level.

One can argue that the default policy is always on the first position
in the list, but that is rather the implementation detail. A detail
which shall not be used within the tool.

The function `xccdf_policy_model_get_policy_by_id' can already take
the NULL attribute and do the right thing.