oval fts: Rewrite the heuristic for pattern matching on path and filepath
This includes rejecting unusable patterns, extracting a fixed prefix
of the path to speed up the search and obtaining a pattern usable for
patial matching to gain additional speedup. The last task is
complicated by the fact that the versions of libpcre prior to 8.00
don't allow certain patterns to be used.

There are several TODOs left:

The first being a usable error reporting.

The second being a stricter pattern checking. The code is there,
sleeping in the comment block, waiting for the people, who would
probably search for their car keys with a regexp, to fix their OVAL
definitions.