Speed up policy_model creation -- lazy creation of policy.
Do not instanciate xccdf_policy for each of the profiles in constructor.
Create a xccdf_policy only upon demand instead.

Given the model_get_policies() function, this can be seen as something
controversial, but in most cases tools use only a single xccdf_policy
thus creating all of the policies is an unnecessary operation.