Verify applied fix by OVAL. Introduce FIXED test result.
This of course needs to have empty oval_agent session,
otherwise it will not evaluate the system, but only query
oval_resultsModel.

There are some discussions about selectivelly reset variables
of the resultsModel, sysModel and caches of probes, but there
is no consensus whether it is possible to do it in a reliable
way.