Improve oscap-xccdf-generate-fix processing
for cases when TestResult is missing.

Previously, we have been using XSLT transformation,
but it is hard to implement Text Substitution in XSLT
and additionally it seems impossible to implement CPE
in XSLT.

This move might be usefull for anaconda, which could
issue oscap command to determine special remeditaion
prescriptions which need to be applied prior the
installation. Example:

    $ oscap xccdf generate guide --profile myprofile \
      --template urn:redhat:anaconda:pre XCCDF.xml