Allow --cpe argument when generating fix from source XCCDF.
For now this option is intentionally left undocumented.
In most cases we still rely on xsl/fix.xslt. Which does
not process CPE. This option shall be documented once
the xsl/fix.xslt is gone.

This option will be used by Anaconda however, to allow
different Anaconda_Fixes for different Anaconda versions.