Allow DataStream when generating fix.
For now this option is intentionally left undocumented.
In most cases we still rely on xsl/fix.xslt. Which does
not work with DataStream. This option shall be documented
once the xsl/fix.xslt is gone.

This option will be used by Anaconda however, to allow
use of Source DataStream Collection during installation.