Avoid exporting duplicate components in the same batch run
We may still export component which is duplicate to the previously
added one, however that is hard to spot.

In future, we may need to have deduplication algorithm, which will
go through the DOM structure of each two components. Noteworthy,
the plain-text comparison seems not be feasible since namespaces
may be named differently and so on.