Add ./run script.
In libguestfs & libvirt we have very successfully use a ./run script
in the top level directory to set environment variables so the program
can be run from the local directory.

eg:
  ./run utils/oscap oval eval [...]
or:
  ./run gdb --args utils/oscap oval eval [...]
or from another project that uses openscap libraries:
  ../openscap/run ./another-program

The same idea appears in the utils/oscap-local.sh script, but ./run is
more flexible since it can also be used for tests, or for running any
external program that happens to use the openscap libraries.

As well as being easier to type.

Signed-off-by: Martin Preisler <mpreisle@redhat.com>