XSLT to remove xccdf:sub elements with broken @idref
Rationale: Official USGCB guidance for RHEL 5 Desktop contains
four <xccdf:sub> elements which refer to nonexistent values.
These 4 elements remain unnoticed as long as the content is
in XCCDF version 1.1. However, the XSD validation schema for
XCCDF 1.2 is more strict and it is picking up on these.

This XSL transformation comes in handy for those, who want
to convert their USGCB to XCCDF 1.2.