Introduce mapping between variables and definitions in definition_model.
Variable Instance Implementation (1/8)

That will come in handy when binding a new value to a variable which
was already assigned with another value (multiset). Such new binding may
require that some of the result_definitions needs to be "probed" again
if evaluated. This mapping helps determine which definitions are dependent
on a given variable.