XML export of multiple tests with various variable_instance attributes
The export of OVAL result is driven by input definitions and directives.
It can happen (given a directive) that there are two result-definitions
for a single definition and only one (be it variable_instance="2" should
be exportted to XML. Then all the result-tests with variable_instance="2"
should be exported along. While, the very same set of result-tests with
variable_instance="1" shall be ommitted from export.