Introduce the variable_instance_hint of collected objects
The hint has meaning that any possible future probe_query of
the given syschar (collected object) neets to create a new
syschar and not re-use the old one. In other words, any next
run of oval_probe_query_*() needs to collect object with the
variable_instance equal to variable_instance_hint.

This is a concept analogous to variable_instance_hint of
oval_result_definition which has been recently introduced.