tests: Expect multiple collected objects for different variable sets
Since the previous changeset, we create new collected objects
and items when evaluating twice; each time with different
variable values.

Note that in this particular test scenario the variable (and its new
values) are bound to a state. That could allow us to re-use existing
syschar. When evaluating the test we could compare the object twice
each time with a different state. In that case we would have only one
set of collected objects/items.

However, we choose to take fresh object/item for each value set. This
will allow us to analyze (oscap oval analyze) more efficiently.