Refactor: Rename property from deprecated to deprecated_by
This is needed because these are two different attributes in the XML
format. Since they are not dependent, we need two separate properties
to keep track of them.

Previous assumption in the code was that @deprecated=false attribute
is equivalent to deprecated property being NULL. That assumption was
proved wrong by official-cpe-dictionary_v2.2.xml.