Prefer lazy linking (LIBADD over LDFLAGS).
This partly reverts 6a6c45fc2d87b8dc7ecbc74db2bd4825df0b0b18,
however it does not break --disable-shared build.

Per automake, LDFLAGS is used early in the line, and LIBADD
is used late. LIBADD is preferred because it seems to make
Make aware of the dependencies and enforces relink any time
it is needed.

Previously, if you made a change in, say, src/common probes
were not relinked and used the old implementation.