Refactor: Rename structure: sds_registry --> sds_session
I was truly struggling to find best name for this structure. It relates
to the fact that I was also struggling to draw line about sds_session
responsibilities.

First I thought it will be something like oscap_source managing only
low level thing of split DS oscap_source to many OVAL/XCCDF
oscap_sources. However, I later found that notion of DataStream
internals needs to be build in to roots of sds_session to be really
useful. That was a time I decide sds_registry is appropriate name.

Later on, I understood that whole registry of DataStrem content is not
needed. And that caller would need to remember datastream_id and lot of
component_ids. That's how the idea of sds_ctx emerged. Then I think it
will be much more intuitive to call this sds_session, since we already
have xccdf_session and there are certain similarities.