Allows for elements like: <plain-text id='empty'/>
The issue manifestated itself, when we started to use xmlReaderWalker.
As a DOM representation of <plain-text id='empty'></plain-text> is
<plain-text id='empty'/>. That is hard to parse.

Let me explain why we use xmlTextReaderReadInnerXml call here. The
problem is that once we move from element to attribute we cannot
easily read the text value of the element. Remember that
xmlTextReaderReadString cannot be used as it does not work with
xmlReaderWalker. The xmlTextReader*Value cannot be used as it will
just return value of the attribute. The xmlTextReaderIsEmptyElement
is not useful either.