Refactor: rename variable.
This needs to be done to improve readability. This whole relative
versus absolute paths is a mess. There are some variables that can
carry only filename others can carry filepath (some relative, some
absolute or some both).

There is a benefit of passing down the relative path. Because relative
path is the catalogue value and the name how the file is known to
other components.