Pass version_info to xccdf_item_to_dom
This allows tailoring export to function properly with oscap_source.
I am not terribly happy about this solution but it doesn't break
behavior and it doesn't break API. Right now the xccdf items don't have
to know their XCCDF version to function properly but that may change in
the future. At which point we will have to deal with this issue in a
cleaner way.