Changed scap-as-rpm to work with Python 2.6
Thanks goes to Tilo Mey for the report and initial patch. I couldn't use
the initial patch as is because it would silently pass when `rpm` is not
available. Or in other cases where `rpm` returns a non-zero exit code.
The function I ended up using instead is by Peter Astrand and Eduardo Felipe,
taken from Python 2.7.