Changed masking behaviour.
According to the clarified specification, the "mask" attribute
should not affect oval_definitions section in oval_results.
It says: ... the"oval_definitions" section must not be altered
and must be an exact copy of the definitions evaluated.
The behaviour has changed since OVAL 5.10,
we must keep the old behaviour for OVAL 5.9 and older.