Enhance README and Convert it into Markdown.
The markdown version of the README document is already in the maint-1.2 but we
also want it in the the maint-1.0 branch so we can have all the contribute
documents and manuals in the maint-1.0. The commit also adds links to the new
"How to contribute to the openscap" and versioning and will replace the old link
to documentation to our new oscap user manual.