Group rules by PCI DSS requirement in HTML report
References to PCI DSS have been added recently for rules in SCAP Security
Guide. Although it is possible now to group these rules by their
PCI DSS identifiers in HTML report, it is not user-friendly.
An URL is displayed in combobox and the groups are in random order.
This commit improves the usability of report. First, "PCI DSS Requirement"
is displayed in combobox instead of an URL. Second, groups are sorted
by the requirement number.