RHBZ#1275369: Fix retrieving SCE results from external results file
OpenSCAP was not able to show SCE results from external SCE results file
in the HTML report, because incorrect XPath expression was in our XSLT
template which is used to genarate the HTML report.
According to the XCCDF specification, the path to the external SCE results file
should be found in a "href" attribute of a "check-content-ref" element.
But we were trying to get it from a "check" element, where it shouldn't be.
This caused I/O warnings. In my opinion there must be some cases when
some information would be missing in a HTML report.