Test behavior of OVAL module with invalid OVAL content
This test tests "oscap oval eval" with "--skip-valid" option
on invalid OVAL content. The tested content references a not
existent state. This situation was one of two reasons of
a segmentation fault, described in issue #191.
The issue #191 has been fixed already. This commit wants to
test it to avoid regressions in future.