Issue #272: XCCDF rule unscored role is not working
A rule with unscored role should not be counted in score.
XCCDF sepcification says about xccdf:Rule/@role="unscored":
if the rule is selected, then check it and include the results in
any report, but do not include the result in score computations.
This commit fixes counting scores to be in line with specification.