Report errors on evaluation of object component of an variable
This commit removes TODOs in _oval_component_evaluate_OBJECTREF() function
by implementing error reporting as required by OVAL specification:
If an entity is not found with a name that matches the value of the item_field
an error is be reported when determining the value of an variable.
Similar statement applies also for entities of the record data type.