Test reporting errors on evaluation of object component of an variable
This test tests whether The ObjectComponentType complex type is processed
as defined in OVAL specification.  It contains one correct case and two
cases which should lead to report an error.
For details see comments in supplied OVAL content.
However the test does not test fields of the record data type, because the record
data type is used only within ldap_item and sql57_item which we don't test in upstream.