Allow both arfvocab and arfrel in core:relationship
To my understanding of NISTIR-7694 is not very clear on the topic.
Reading the prose, the 'isAbout' is supposed to be a part of arfrel,
however in the example, the isAbout is prefixed by arfvocab.

I guess it is better to suppose both.

Afterall, so far, we are exporting arfrel:isAbout, while we are
supporting only arfvocab:isAbout when reading it.