Add a message about filter used within an object
OVAL objects can contain optional filter element. The filter
references a state. The filter can exclude or include items
conforming to the state from the set of collected objects.
Since the filter can significantly affect the result of
object evaluation it would be nice to inform user about
processing the filter. This commit adds a short message to the
verbose log.