Voodoo: Tell us where is zombie
We think she left dead, but the priest (ps) conceals where it is.

This is not a black magic really. We think the zombie process should be
listed in cemetery, but sometimes it is missing. Good thing is to put
more details into the log. Next time we need to record the moon phase.

See github issue #315 for more details.