Add message about extend definition
In OVAL the definitions usually contain one or more tests,
but they can also be extended by referencing to other definitions.
The refrenced definitions are required to be evaluated to get
final result of the definition under the evaluation.
This construct is sometimes used in SSG.
Currently we don't show the information about extending definition
in verbose log. The log is very confusing for a reader if a
extended definition is evaluated.
This commit adds an info message when extend_definition is used.