Add a pointer to result system into probe session
We would like to move implementation of evaluating results
to oval_probe_query_definition().
To do that, we need to access result system from this function.
But it is not possible to do now, because the function consumes
only the oval_probe_session.
To access the result system from there, we need to add the result
system into the probe session.