Change oval_probe_query_criterion to oval_probe_query_test
This function takes oval_criteria_node *cnode as its second paramater.
But it assumes that cnode is a criterion and it only uses this paramater
to get the oval_test referenced by the criterion.
The function is static and it is called only in one place,
where we are already sure that cnode is a criterion.
So we can move some part of code to the caller in a way
that we can change the type of second parameter to oval_test.
This change is done because we would like to reuse the code
in results_model.