Describe "--sce-template" option in man page
The "--sce-template" option is available in oscap xccdf generate report
module for a long time (since f88d6cee276ef56464e591629e484041864e0f68),
but it was not documented in the manual page.
The behavior of this option is very similar to "--oval-template",
expect this one is for SCE results.