Issue#394: Add missing default template for SCE results
The oscap xccdf generate report module is able to embed SCE results
into the HTML report. To do that, a template for results file name
must be specified using the "--sce-template" option.
However, users expect that if they omit "--sce-template" option,
a default file name template will be used.
It works like that for OVAL results, so it should work the same way
also for SCE results.
To fix this issue, this commit adds a default file name template
for SCE into the xccdf module.