Remove callback on RPM < 4.6
Recently we have merged a fix to compile the rpm probes
on maint-1.0 on systems with older versions of rpmlib.
However, the reporting function  is not present there,
so it does not make any sense to set the logging callback,
because it has no effect. This commit removes the callback
on those systems.