Issue #281: Fix invalid characters in remediation results
We include stdout of a remediation script to the XCCDF results.
But the output of the remediation script can contain characters
that are forbidden in XML documents, eg. most of the non-printable
characters. We need to escape or remove them from the string
before we include it to the XML document. Libxml provides a
function to get rid of those characters.