Adjust the valgrind test to detect all kinds of memory errors
This commit enables us to detect also other memory errors,
not only the memory leaks, but also invalid reads etc.
We will use --error-code to detect that valgrind failed.
This option is designed for usage in automated test suites (see man).
Parsing the log is not necessary to detect the fail.
Also valgrind has to generate a special logfile for each child process
to avoid undefined results (see man).