Fix export of overrided oscap_text
This change allows attributes marked as overriders by
oscap_text_set_override() to be exported.

Exportation of an overrider attribute is conditioned by override_given.
But override_given is only set when an imported attribute has override
capability set. So, when an oscap_text was marked as overrider by
oscap_text_set_set_override() the attribute was never exported.

Related to RHBZ#1320194.