Use ARF report ID while creating ARF report
Instead of XCCDF Benchmark ID, we will put the real OVAL results ID
as ARF report ID. This will allow us to reference the part of
ARF report conataining OVAL results. We will use this ID in
xccdf:rule-result/check-content-ref@href attribute.
We need to figure out which OVAL results corresponds with
which OVAL definitions document, so we use the mapping hash tables
we created and passed in previous commits.