Introduce a test for XCCDF 1.1 to XCCDF 1.2 transformation
We ship an XSLT template that allows us to convert older
XCCDF files in XCCDF 1.1 to XCCDF 1.2. The template can be found
in xsl/xccdf_1.1_to_1.2.xsl. However we don't test the transformation.
Since the transforamtion is mentioned in user manual, we should
have a test that ensures it is working. This commit introduces
a simple test based on XCCDF content from NIST test suite.