Do not rely on number of lines on stderr
We fixed duplicate error messages (issue #707). But in this test,
we assumed that there are 6 lines on stderr. Actually two of them
were duplicated. We should not count number of lines on stderr
because we may add a new warning somewhere in our code and it will
break the test again. This commit stops testing stderr line count.