Issue #723: Don't require other resources in "generate fix"
Checks are not necessary to load OVAL because the remediation
scripts are stored in the XCCDF. This will remove the warning
for remote resources that are not needed. The commit introduces
new type, xccdf_session_flags_t, which speciefies flags
for XCCDF session. These flags drive loading content in XCCDF
session.