Fix Python 2 Bindings (swig)
We have to be explicit on Python version, otherwise we'll get
a horrible mix of Python 2 and 3. We also need to link all libraries
with _openscap_py.so to avoid error messages about undefined symbols.
We also need to specify module name as "openscap_py", beacuse otherwise
python will complain that function "init_openscap_py" isn't defined
in _openscap_py.so.