Replace more breaks by gotos
The break statement exits the for loop. The SKIP_LOOP label is placed right
after the end of the for loop. Therefore it is possible to replace this break
by goto SKIP_LOOP. It will help us in future when we will replace labels in this
for loop by a switch, the break has a different meaning inside a switch.
This is very similar to the previous commit.