Add a function that lists available probes
This function will be used in 'oscap --version'. To use any function
in the oscap utilty the function has to be a part of public API. Therefore
we need to move the header file to a public directory. We also need to
replace undescore in the file name to prevent ignoring by .gitignore,
because .gitignore ignores built probe binaries that start by 'probe_'
prefix.