Terminate probe_input_handler_thread
In past, the probe process had a signal handler, which canceled
the probe_input_handler thread when the process received SIGINT.
However, we want to get rid of signals due to Windows support
and we also don't want to send SIGINT by 'oscap' process to
himself. In sch_queue_close we cancel the probe_common_main thread,
so we should also terminate the probe_input_handler_thread. This
can be done in a cleanup routine, which is trigerred by cancelation
of the probe_common_main_thread. This code is inspired by original
signal_handler thread code from maint-1.2 branch.