Do not convert SEXP to string and string back to SEXP
Both sides of communication use SEXP. One side wants to send a SEXP,
the second side wants to receive a SEXP. If they could pass a pointer
to SEXP to each other directly, the sender will not need to convert SEXP
to string, allocate memory for that string and the receiver will not need
to parse the SEXP from string back. We can simplify it a lot.