Do not setup verbose mode twice
When the probe was a separate process, we had to pass it the verbosity
level and log file name via 2 environment variables. Also, all the
probe processes had to open and close the log file on its own.
Now, when avarything is only a single process, we should setup
the verbose mode only once. Also, we don't need the environemt
variables. It also prevents repeated close()s on the same file.