Remove unneeed header file _seap-scheme.h
SCH_QUEUE is the only definition that is used in OpenSCAP, but it is used
in only one file (seap.c), so we can move the definition of SCH_QUEUE
to seap.c and then we can remove the _seap-scheme.h and all the references
to it. It will simplify our code, because this way we will finally
get rid of possibility of having multiple SEAP schemes.