Fix SEXP datatype in xmlfilecontent probe
The tests/mitre/test_mitre_unix_probes.sh test expects at XPath values
a datatype="int" attribute on ind-sys:xmlfilecontent_item/ind-sys:value_of
element. However the probe produced a float number. This worked before.
When we converted SEXP to string and then back to SEXP, it translated
to the same string and the SEXP parser then recognized the value as integer.
However, now, when we pass the SEXP directly to the OVAL results, the bug
is not hidden anymore. We need to produce a correct datatype in probe.