Change textfilecontent(54) probes to not use chroot
Adds a new parameter, 'prefix' to process file and to oval_fts_open.
This prefix is added to the FTS paths, and then removed from FTS paths,
and the added to lstat(), and open() calls. The prefix should not appear
in the OVAL results document, so that the user will feel like it has
chroot()ed. But chroot() is not used at all in these probes anymore.