Add IP addresses in XCCDF results on Windows
I found this code in my git stash, so I think it would be useful
to create a commit from it and share it on Github.
There is a second place where we collect IP addresses of the
scanned machine. These are saved to XCCDF results and are shown
in HTML report. So this change is very visible.
I understand that XCCDF and OVAL are different things, however
I don't like that this copies a lot of code from the system info
probe.