Fix coversion of SEXP to struct oval_record_field
The member name of struct oval_record_field should contain
the value of 'name' attribute of the OVAL 'field' element.
However, it was populated by name of the element, which is
always 'field'. The values were present in SEXP, but due to
this bug, the value of name attribute got lost at this place.
Before:
<field name="field">Brno</field>
After:
<field name="city">Brno</field>