Fix oscap_dirname again
The condition was wrong, it should require only one of forward
and back slashes to be present, not both of them.
Also, we should not omit the drive letter and add it at the
beginning of the path.
Also, we need to ensure that the returned path does not end with
forward or backward slash to be consistent with Linux dirname()
function.