Change utilities options to cmake_dependent_options
The utilities that do not work on Windows will be disbaled on Windows
by default, but they will be enabled by default on non-Windows systems.
The users building on Windows will not have to turn off these options
manually by CMake command line. This now affects oscap-docker,
scap-as-rpm, oscap-vm, oscap-ssh and oscap-chroot.