Do not use assert in unit tests
Macro `assert` compiles to empty command if build type is Release.
That caused some of the tests failing in Release configuration.
We need to replace `assert` by a macro that is not dependent on
NDEBUG variable. This commit introduces `oscap_assert` and replaces
all occurences of `assert` in tests by `oscap_assert`.
Fixes #912.