Add libprocps-dev to Travis CI and Ubuntu dependencies
libprocps-dev provides `proc/devname.h`, which provides `dev_to_tty`
function, which is used by process probe and process58 probe.
It is better to use the system library version than the custom
version from `compat` if the system version is available.
The Travis CI will now install this package and we will recommend
users to do that as well.