Install bash completion to a right directory
Using CMAKE_INSTALL_SYSCONFDIR caused that the "bash completion"
file was installed into `/usr/etc`, but it should go to '`/etc`.
To fix that, we need to use a special case described in
https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html:
For <dir> equal to SYSCONFDIR, LOCALSTATEDIR or RUNSTATEDIR,
the CMAKE_INSTALL_FULL_<dir> is computed by prepending just / to
the value of CMAKE_INSTALL_<dir> if it is not user-specified as
an absolute path.
That means we should use CMAKE_INSTALL_FULL_SYSCONFDIR instead
of CMAKE_INSTALL_SYSCONFDIR as a install destination.