Change CMakeLists to be able to use asciidoc
Asciidoc could be used instead of asciidoctor. Changed the command line
to be compatible with both programs. Explicitely set the HTML5 backend
because it is not default in asciidoc. We have to make the target
directory before we call the program, because asciidoc doesn't create
directories automatically. If both asciidoc and asciidoctor are found,
asciidoc is preferred.