Clean up implementation of strto_ functions
Because strtoll/strtoull take a null-terminated
C-string, keep the strto_ functions as wrappers,
but explicitly null-terminate inputs. This is useful
in the parser, where inputs to strto_* are length
terminated substrs of the input.

Fixes the SEAP parser test case failure on gcc 8.1.1
by removing the internal implementation and using strtoll
instead.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>