Remove option '--show' from 'oscap xccdf generate report'
This option has been replaced by a possibility to hide/show selected
XCCDF result types by clicking on checkboxes in the HTML report.
Moreover, this option does not work now. I think it is a leftover
from the old style report, which was kept due to backwards compatibility.
With upcoming major release we can remove it.