Refactor: Add a wrapper function around probe_table_list
The function 'probe_table_list' is the only function from
probe-table.h that needs to be public, because it is used from
'oscap' utility to print the list probes using 'oscap --version'
command. Wrapping it in a public API wrapper will allow us to
move the probe table API to a private header file.