Skip sysctl probe test if no SysV services are found
If the list of services returned by `chkconfig --list` is empty,
then the generated OVAL file would be invalid. Since there would
not be anything to test in that situation, it would be better
to skip the test instead. This situation can happen on modern
Linux distributions that have only native systemd services.