Fix broken compiler visibility presets
The main problem is that C is not C++, therefore the symbol visibility
setting for C++ is not applicable to C. This caused that many symbols
which must be private were actually exported in libopenscap.so and they
were publicly visible. I have discovered this implication of C++
not being C by using `nm` command and analysing `nm` output carefully.
Fixes #963.