Move public symbols to a public header
Our header file src/common/util.h deifned both public and private
symbols. If they are used in probes, they can be made private.
Previously the probes used to be separated binaries, so all the
functions used in probes had to be made public. However, this is
no longer true, so these comments are not relevant anymore and they
should be removed. On the other hand, if some of the functions is
used in oscap utility, they need to be kept public. This commit
moves the symbols that need to be kept public to a new header file
src/common/oscap_helpers.h Moreover, the symbols that doesn't have
to be public won't be public anymore.