Do not use dynamic symbols on OS X
The way of using export symbols list is different on OS X,
therefore we need to amend the generator script and
restrict the usage of symbols list ot Linux only.
In future, we plan to add a symilar mechanism on OS X as well,
but we need to clean up the header files first.