Add a workaround for invalid Windows 7 CPE
Official Windows 7 CPE according to National Vulnerability Database
CPE Dictionary as of 2018-08-29 is 'cpe:/o:microsoft:windows_7'.
However, content exported from Microsoft Security Compliance Manager
as of version 4.0.0.1 in CAB archive using 'Export in SCAP 1.0' is
'cpe:/o:microsoft:windows7'. If this pattern is matched, we will add
an underscore to workaround the situation that this XCCDF benchmark is
not applicable.