Don't free oscap_source when parsing cvrf_xml
Do not free oscap_source when parsing cvrf_xml for index.
It should not be the duty of cvrf xml parsing function to free its
source of input.

Plus, cvrf_index_import is also called in function
cvrf_session_new_from_source. In cvrf_session structure,
the import source is part of the strucutre along with cvrf index, it
make no sense to free it.