tests/probes/sysctl/test_sysctl_probe_all.sh: updated for procps-ng > 3.3.12
Some sysctl options cannot be read until they are set. Sysctl from
procps-ng > 3.3.12 reports an error when reading these options but still
prints them to stdout. On the other hand openscap does not include these
options into results when they are not set. This was introduced in the commit
7b8ca9c and it is causing different results between openscap and sysctl from
procps-ng > 3.3.12. This commit updates the sysctl test to count with these
differences in case that procps-ng > 3.3.12 is installed on the system.