Check if system has enough memory
As a team effort we tried to mitigate `oscap` memory consumption.
We found out that there is already code that checks for memory
limits, but unfortunately it was never run, because a wrong SEXP
was measured. We will instead count items collected by each thread
and every 100 collections we will check if the system has enough
free memory.