Show details in HTML report by default if --results-arf is given
This commit changes the `oscap` behavior in the following way:
If `oscap xccdf eval` is called with both `--results-arf` and
`--report` the HTML report will contain `OVAL details` section
even if `--oval-results` is not given by the user.
I think we can do it because the ARF file already contains OVAL
results anyway. I think it's not convenient to type `--oval-results`
explicitly.